QPID-8144: [Broker-J] Correct arithmetic type used when computing utilised direct memory (from int to long).
Reimplementation of equivilent change from master.

+++ common/src/main/java/org/apache/qpid/bytebuffer/QpidByteBuffer.java
@@ -835,7 +835,7 @@ public class QpidByteBuffer
 
     public static long getAllocatedDirectMemorySize()
     {
-        return _pooledBufferSize * getNumberOfActivePooledBuffers();
+        return (long) _pooledBufferSize * getNumberOfActivePooledBuffers();
     }
 
     public static int getNumberOfActivePooledBuffers()
